Category delete case added

+++ src/controllers/categories.ts
@@ -1,7 +1,8 @@
 import type { RequestHandler } from 'express';
 import type { z } from 'zod/v4';
+import type { UserType } from '#types';
 import type { categoryInputSchema, categorySchema } from '#schemas';
-import { Category } from '#models';
+import { Category, Product } from '#models';
 
 type categoryInputDTO = z.infer<typeof categoryInputSchema>;
 type categoryDTO = z.infer<typeof categorySchema>;
@@ -42,6 +43,10 @@ export const deleteCategory: RequestHandler<{ id: string }> = async (req, res) =
   const {
     params: { id }
   } = req;
+  const products = await Product.find({
+    categoryId: id
+  });
+  if (products.length > 0) throw new Error('Category cannot be deleted because it has products', { cause: 409 });
   const category = await Category.findByIdAndDelete(id);
   if (!category) throw new Error('Category not found', { cause: 404 });
   res.json({ message: 'Category deleted' });
